+++ config/filament-seo-meta-tags.php
@@ -1,0 +1,6 @@
+<?php
+
+// config for MartinRo/FilamentSeoMetaTags
+return [
+
+];